Crate print

Allows crates in the kernel to log messages to the default terminal using the println! and print! macros

The current system simply enqueus the Print event into the print queue of the default terminal

Macros

print

The main printing macro, which simply pushes an output event to the event queue. This ensures that only one thread (e.g., a terminal acting as a consumer) ever accesses the GUI.

println

Calls print!() with an extra newilne \n appended to the end.

Functions

print_to_default_output

Enqueues the given fmt_args as a String onto the default printing output queue, which is typically the default terminal application’s input queue

set_default_print_output

Gives the kernel an endpoint (queue producer) to which it can send messages to be printed